api/requirements.txt
asgiref==3.3.1
Django==3.1.7
djangorestframework==3.12.2
Pillow==8.1.2
pytz==2021.1
sqlparse==0.4.1
absl-py==0.11.0
astor==0.8.1
backcall==0.2.0
bleach==1.5.0
cached-property==1.5.2
certifi==2020.12.5
colorama==0.4.4
cupy-cuda100==7.1.1
cycler==0.10.0
decorator==4.4.2
easydict==1.9
fastrlock==0.5
gast==0.2.2
google-pasta==0.2.0
grpcio==1.36.1
h5py==2.10.0
html5lib==0.9999999
imageio==2.9.0
importlib-metadata==3.7.0
ipykernel==5.5.0
ipython==7.16.1
ipython-genutils==0.2.0
jedi==0.18.0
jupyter-client==6.1.11
jupyter-core==4.7.1
Keras==2.2.4
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
Markdown==3.3.4
matplotlib==3.3.2
networkx==2.5
numpy==1.18.5
opt-einsum==3.3.0
parso==0.8.1
pickleshare==0.7.5
Pillow==7.2.0
prompt-toolkit==3.0.16
protobuf==3.15.4
Pygments==2.8.0
pyparsing==2.4.7
python-dateutil==2.8.1
PyWavelets==1.1.1
pywin32==300
PyYAML==5.4.1
pyzmq==22.0.3
scikit-image==0.17.2
scipy==1.5.4
six==1.15.0
tensorboard==1.8.0
tensorflow-estimator==1.15.1
tensorflow-gpu==1.8.0
termcolor==1.1.0
tifffile==2020.9.3
tornado==6.1
tqdm==4.49.0
traitlets==4.3.3
typing-extensions==3.7.4.3
wcwidth==0.2.5
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.4.0
